mqtt_bridge/requirements.txt
paho-mqtt>=1.6.1
influxdb-client>=1.33.0
requests>=2.28.0
